two post-November bug fixes: lat/lon typo, and checkpoint finder
The checkpoint system previously required the narrator to pass within a five-minute walk of each one to tick it off; this meant it was possible for them to miss a checkpoint entirely if they were travelling at speed. (There was some old code that caught such passes and threw the narrator back, but it was dropped when I changed the wayfinding system and I forgot to adjust the checkpoint check.) Also fixed a lat/lon typo for detecting the 180th meridian (which possibly didn't matter previously, but is needed to get the new checkpoint check right) and switched off some HTML-comment debug code.